Test passes for edge
